third_party/blink/web_tests/external/wpt/html/dom/idlharness.https_include=HTML._-expected.txt

This is a testharness.js-based test.
Found 16 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] HTMLElement interface: attribute accessKeyLabel
  assert_true: The prototype object must have a property "accessKeyLabel" expected true got false
[FAIL] HTMLElement interface: document.createElement("noscript") must inherit property "accessKeyLabel" with the proper type
  assert_inherits: property "accessKeyLabel" not found in prototype chain
[FAIL] HTMLMediaElement interface: document.createElement("video") must inherit property "fastSeek(double)" with the proper type
  assert_inherits: property "fastSeek" not found in prototype chain
[FAIL] HTMLMediaElement interface: calling fastSeek(double) on document.createElement("video") with too few arguments must throw TypeError
  assert_inherits: property "fastSeek" not found in prototype chain
[FAIL] HTMLMediaElement interface: document.createElement("video") must inherit property "getStartDate()" with the proper type
  assert_inherits: property "getStartDate" not found in prototype chain
[FAIL] HTMLMediaElement interface: document.createElement("audio") must inherit property "fastSeek(double)" with the proper type
  assert_inherits: property "fastSeek" not found in prototype chain
[FAIL] HTMLMediaElement interface: calling fastSeek(double) on document.createElement("audio") with too few arguments must throw TypeError
  assert_inherits: property "fastSeek" not found in prototype chain
[FAIL] HTMLMediaElement interface: document.createElement("audio") must inherit property "getStartDate()" with the proper type
  assert_inherits: property "getStartDate" not found in prototype chain
[FAIL] HTMLMediaElement interface: new Audio() must inherit property "fastSeek(double)" with the proper type
  assert_inherits: property "fastSeek" not found in prototype chain
[FAIL] HTMLMediaElement interface: calling fastSeek(double) on new Audio() with too few arguments must throw TypeError
  assert_inherits: property "fastSeek" not found in prototype chain
[FAIL] HTMLMediaElement interface: new Audio() must inherit property "getStartDate()" with the proper type
  assert_inherits: property "getStartDate" not found in prototype chain
[FAIL] HTMLMediaElement interface: operation fastSeek(double)
  assert_own_property: interface prototype object missing non-static operation expected property "fastSeek" missing
[FAIL] HTMLMediaElement interface: operation getStartDate()
  assert_own_property: interface prototype object missing non-static operation expected property "getStartDate" missing
[FAIL] HTMLTemplateElement interface: attribute shadowRootDelegatesFocus
  assert_true: The prototype object must have a property "shadowRootDelegatesFocus" expected true got false
[FAIL] HTMLTemplateElement interface: document.createElement("template") must inherit property "shadowRootMode" with the proper type
  assert_equals: expected "string" but got "object"
[FAIL] HTMLTemplateElement interface: document.createElement("template") must inherit property "shadowRootDelegatesFocus" with the proper type
  assert_inherits: property "shadowRootDelegatesFocus" not found in prototype chain
Harness: the test ran to completion.
